mild restructuring
